requirements.txt
streamlit
scikit-learn
numpy
pandas
python-docx
reportlab
